Fix review comment: not reasons no check comparator compatibility
In C++17 this is stricter than the algorithm and the SIMD body: is_invocable_r_v<bool> requires an implicit conversion to bool, while Compare and these if/! expressions only require contextual conversion. A valid result type with explicit operator bool() therefore compiles in the SIMD expressions but makes this gate select the serial fallback. Detect invocability plus the actual contextual/negation expressions instead so conforming C++17 comparators are not unnecessarily de-vectorized, and add a compile-time case for that result type.
